Added many pages, some navigation work, some registration page work
Added the following pages:
- Student settings
- Student profile
- Shadowing offer detail
- My applications (shows student all offers they are currently applied to)
- (renamed) Student main
- Business main
- Business settings
- Business profile
- Edit/create/remove offer
- Offer applications (shows everyone that is applied to an offer)

Registration page adaptability (fields changing depending on whether the person signing up is a student or a business representative) is now fleshed out and just needs to be filled in with specific settings.

Almost all student navigation is ready. Business navigation is not yet there at all.