Added RecyclerView offer selection, Business/Student distinction, further navigation, renamed some files
Clicking an item in the RecyclerView now takes you to a new activity that knows which item was selected.
The login now sends the user to a different page depending on whether they are registered as a student or a business.
The business main page now has functional navigation to the settings and profile pages, as well as a functioning logout button.
Renamed some files for clarity and consistency.